+++ libherokubuildpack/src/inventory/unit.rs
@@ -1,0 +1,11 @@
+use crate::inventory::checksum::Digest;
+
+impl Digest for () {
+    fn name_compatible(_: &str) -> bool {
+        true
+    }
+
+    fn length_compatible(_: usize) -> bool {
+        true
+    }
+}