+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `libherokubuildpack`:
+  - Added `inventory` module. ([#861](https://github.com/heroku/libcnb.rs/pull/861))
 
 ## [0.23.0] - 2024-08-28